Fix leftover paths after restructure
