+++ b2b/web-app/components/sections/sections/settingsSection/mfaSection/mfaProviderUtils.ts
@@ -17,9 +17,9 @@
  */
 
 import { EMAIL, EMPTY_STRING, SMS, TOTP } from "@pet-management-webapp/shared/util/util-common";
-import emailOTPImage from "../../../../../../../libs/business-admin-app/ui/ui-assets/src/lib/images/email-solid.svg";
-import smsOTPImage from "../../../../../../../libs/business-admin-app/ui/ui-assets/src/lib/images/sms-otp.svg";
-import totpOTPImage from "../../../../../../../libs/business-admin-app/ui/ui-assets/src/lib/images/totp.svg";
+import emailOTPImage from "../../../../../src/business-admin-app/ui/ui-assets/src/lib/images/email-solid.svg";
+import smsOTPImage from "../../../../../src/business-admin-app/ui/ui-assets/src/lib/images/sms-otp.svg";
+import totpOTPImage from "../../../../../src/business-admin-app/ui/ui-assets/src/lib/images/totp.svg";
 
 export function getImageForMFAProvider(providerId: string): string {
     if (providerId === EMAIL) {
@@ -32,5 +32,4 @@ export function getImageForMFAProvider(providerId: string): string {
         return totpOTPImage;
     }
     
-    return EMPTY_STRING;
-}
+    return EMPTY_STRING;}
