+++ b2b/web-app/src/business-admin-app/util/util-application-config-util/jest.config.ts
@@ -24,5 +24,5 @@ export default {
     "^.+\\.[tj]sx?$": ["@swc/jest", { jsc: { transform: { react: { runtime: "automatic" } } } }]
   },
   moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
-  coverageDirectory: "../../../../coverage/libs/business-admin-app/util/util-application-config-util"
+  coverageDirectory: "../../../../coverage/src/business-admin-app/util/util-application-config-util"
 };
